+++ cards.py
@@ -2,26 +2,29 @@ from collections import namedtuple
 
 
 class Card(namedtuple('CardData', ['suit', 'value'])):
-    """This is a class for all cards in a deck displayed in tuples.
-
-    Constructor:
-    def __new__(cls, suit, value):
-        :param cls: Card class reference.
-        :param suit: The suit of the card instance.
-        :param value: The value of the card instance.
-        :return: An instance of card object with specified suit and value.
+    """This is a class for all cards in a deck.
 
     Attributes:
-    suit -- The suit of the card.
-    value -- The value of the card.
+        suit -- The suit of the card.
+        value -- The value of the card.
 
     Methods:
-    __str__ -- Description of class attributes.
+        __str__ -- Description of class attributes.
     """
     suits = '♣ ♢ ♡ ♠'.split()
     values = '2 3 4 5 6 7 8 9 10 J Q K A'.split()
 
     def __new__(cls, suit, value):
+        """Constructor for the Card class.
+
+        Parameters:
+            cls -- Card class reference.
+            suit -- The suit of the card instance.
+            value -- The value of the card instance.
+
+        Returns:
+            An instance of card object with specified suit and value.
+        """
         assert suit in cls.suits
         assert value in cls.values
         return super().__new__(cls, suit, value)